+++ Sources/SwiftHttp/UrlSessionHttpClient.swift
@@ -6,9 +6,6 @@
 //
 
 import Foundation
-#if canImport(FoundationNetworking)
-import FoundationNetworking
-#endif
 
 /// Default URLSession based implementation of the HttpClient protocol
 public struct UrlSessionHttpClient: HttpClient {
@@ -41,28 +38,20 @@ public struct UrlSessionHttpClient: HttpClient {
         if log {
             print(urlRequest.curlString)
         }
+
         let res: (Data, URLResponse)
-        
         let cache: CachedURLResponse? = URLCache.shared.cachedResponse(for: urlRequest)
+
         if (cache != nil) {
-            do {
-                return try HttpRawResponse((cache!.data, cache!.response))
-            } catch {
-                print(error)
-            }
-        }
-        
-#if os(Linux)
-        res = try await asyncMethod(with: urlRequest, session.dataTask)
-#else
-        if #available(iOS 15.0, tvOS 15.0, watchOS 8.0, macOS 12.0, *) {
-            res = try await session.data(for: urlRequest)
+            // Set res to cached reponse and data
+            res = (cache!.data, cache!.response)
         } else {
-            res = try await asyncMethod(with: urlRequest, session.dataTask)
+            // Send request
+            res = try await session.data(for: urlRequest)
+            // Save response and data
+            URLCache.shared.storeCachedResponse(CachedURLResponse(response: res.1, data: res.0), for: urlRequest)
         }
-#endif
-        
-        URLCache.shared.storeCachedResponse(CachedURLResponse(response: res.1, data: res.0), for: urlRequest)
+
         return try HttpRawResponse(res)
     }
     
@@ -83,21 +72,20 @@ public struct UrlSessionHttpClient: HttpClient {
         if log {
             print(urlRequest.curlString)
         }
+
         let res: (Data, URLResponse)
+        let cache: CachedURLResponse? = URLCache.shared.cachedResponse(for: urlRequest)
         
-#if os(Linux)
-        res = try await asyncMethod(with: urlRequest) {
-            session.uploadTask(with: $0, from: data, completionHandler: $1)
-        }
-#else
-        if #available(iOS 15.0, tvOS 15.0, watchOS 8.0, macOS 12.0, *) {
-            res = try await session.upload(for: urlRequest, from: data, delegate: nil)
+        if (cache != nil) {
+            // Set res to cached reponse and data
+            res = (cache!.data, cache!.response)
         } else {
-            res = try await asyncMethod(with: urlRequest) {
-                session.uploadTask(with: $0, from: data, completionHandler: $1)
-            }
+            // Send upload request
+            res = try await session.upload(for: urlRequest, from: data, delegate: nil)
+            // Save response and data
+            URLCache.shared.storeCachedResponse(CachedURLResponse(response: res.1, data: res.0), for: urlRequest)
         }
-#endif
+
         return try HttpRawResponse(res)
     }
     
@@ -115,32 +103,12 @@ public struct UrlSessionHttpClient: HttpClient {
         if log {
             print(urlRequest.curlString)
         }
-        let res: (URL, URLResponse)
-#if os(Linux)
-        res = try await asyncMethod(with: urlRequest, session.downloadTask)
-#else
-        if #available(iOS 15.0, tvOS 15.0, watchOS 8.0, macOS 12.0, *) {
-            res = try await session.download(for: urlRequest, delegate: nil)
-        } else {
-            res = try await asyncMethod(with: urlRequest, session.downloadTask)
-        }
-#endif
+
+        let res: (URL, URLResponse) = try await session.download(for: urlRequest, delegate: nil)
+
         guard let pathData = res.0.path.data(using: .utf8) else {
             throw HttpError.invalidResponse
         }
         return try HttpRawResponse((pathData, res.1))
-    }
-    
-    private func asyncMethod<T, S: URLSessionTask>(with urlRequest: URLRequest, _ method: @escaping (URLRequest, @escaping @Sendable (T?, URLResponse?, Error?) -> Void) -> S) async throws -> (T, URLResponse) {
-        try await withCheckedThrowingContinuation { continuation in
-            method(urlRequest) { t, response, error in
-                if let t = t, let response = response {
-                    continuation.resume(returning: (t, response))
-                } else {
-                    continuation.resume(throwing: error ?? HttpError.invalidResponse)
-                }
-            }
-            .resume()
-        }
     }
 }